Procedural Dungeon Start
Got the basics down for the Procedural Dungeon Mechanic.

Created a fail safe for if the dungeon is taking to long to load - it will re generate the dungeon if it takes longer than 8 seconds to load.

Also Created a hole blocking system to block unwanted holes at the end of each room when the dungeon finishes generating.